==> 01_get_metadata.Rmd <==
---
title: "Get FatT transcript metadata"
author: "Nathaniel Olin"
date: "`r date()`"
output: github_document
editor_options: 
  chunk_output_type: console
---

```{r setup}
library(tidyverse)
library(googlesheets4)
```

# Get list of episodes

```{r get-list}
gs4_deauth() # Public sheet does not require auth

seasons <- list(
  "Autumn in Hieron",
  "Marielda",
  "Winter in Hieron",
  "Spring in Hieron",
  "COUNTER/Weight",
  "Twilight Mirage",
  "Road to PARTIZAN",
  "PARTIZAN"
)

dat <- lapply(
  seasons,
  function(x)
    read_sheet(
      "1KZHwlSBvHtWStN4vTxOTrpv4Dp9WQrulwMCRocXeYcQ",
      sheet = x
    )
)

# Stack sheets
names(dat) <- seasons
dat <- bind_rows(dat, .id = "season")
dat <- janitor::clean_names(dat)

# Check consistent format of dataframe
stopifnot(names(dat) == c("season", "episode", "complete", "doc"))
dat$complete <- dat$complete == "Yes" & !is.na(dat$complete)
```

# Clean List

```{r clean}
remove_print <- function(d, remove) {
  d %>% 
    filter(str_detect(episode, remove)) %>% 
    select(episode) %>% 
    print(n = Inf)
  d %>% filter(! str_detect(episode, remove))
}

dat <- remove_print(dat, "[Mm]ortem")
dat <- remove_print(dat, "Bonus")
dat <- remove_print(dat, "Introduction|A Message From|A Quick Announcement")
dat <- remove_print(dat, "The Months of Autumn")
dat <- remove_print(dat, "Takin’ a Snow Day")
dat <- remove_print(dat, "Patreon Announcement")
dat <- remove_print(dat, "Gen Con")
dat <- remove_print(dat, "Announcing: PARTIZAN")
```

## Code season numbers

```{r}
dat <- dat %>%
  mutate(
    season_id = recode(
      season,
      "Autumn in Hieron" = 1,
      "COUNTER/Weight" = 2,
      "Marielda" = 2.5,
      "Winter in Hieron" = 3,
      "Twilight Mirage" = 4,
      "Spring in Hieron" = 5,
      "Road to PARTIZAN" = 5.5,
      "PARTIZAN" = 6
    )
  )
```

## Code episode numbers

```{r ep-nums}
# Consistent format
dat <- dat %>%
  mutate(episode = if_else(
    episode == "PARTIZAN 09 - DESERT SQUIRE - VANTAGE",
    "PARTIZAN 09: DESERT SQUIRE - VANTAGE",
    episode
  ))


dat <- dat %>%
  mutate(
    episode_id = as.numeric(str_replace_all(
      episode, "^(.*? )([0-9-]*):.*", "\\2"))
  )

# Holiday episodes
dat %>% filter(str_detect(episode, "Holiday"))
dat <- dat %>%
  mutate(episode_id = if_else(
    str_detect(episode, "Holiday"),
    as.numeric(NA),
    episode_id
  ))
```

## Code filename (for later)

```{r}
dat <- dat %>%
  mutate(filename = sprintf(
    "%s.txt", 
    str_replace_all(
      tolower(episode), 
      pattern = c(" " = "_", "[:!?'’(),./]" = "", "\"" = ""))))
```

# Write out

```{r write}
dat %>%
  select(filename, season_id, season, episode_id, episode, doc, complete) %>%
  print() %>%
  write_csv("meta.csv")
```
