==> 02_get_transcripts.Rmd <==
---
title: "Get FatT transcripts"
author: "Nathaniel Olin"
date: "`r date()`"
output: github_document
editor_options: 
  chunk_output_type: console
---

```{r setup}
library(tidyverse)
library(googledrive)
drive_deauth() # Public sheet does not require auth

if(! "transcripts" %in% list.files())
  dir.create("transcripts")
```

# Download transcripts

## Get episode URLs

```{r get-list}
dat <- read_csv("meta.csv") %>%
  filter(complete)
```

## Save transcripts to file

```{r scrape}
for(i in seq_along(dat$doc)){
  if(! dat$filename[i] %in% list.files("transcripts") & 
     !is.na(dat$doc[i]))
    drive_download(
      as_id(dat$doc[i]), 
      path = paste0("transcripts/", dat$filename[i])
    )
}
```
